[osr] Add new osr-related fields and accessors to FeedbackVector
This adds the OSR state to the FV, where the state consists of the
osr_urgency (same semantics as previously on the BytecodeArray) and a
maybe_has_optimized_osr_code bit (set if any optimized OSR Turbofan code
exists for this function).

The two are packed into one field for efficient OSR checks in generated
code (to be implemented in the followup CL).

Bug: v8:12161
Change-Id: Id4edb8f5db0bf02e0d04b87aeec8d8c53e213503