heap: Inline GCTracer::Scope::Name
This is a follow-up to https://crrev.com/c/3581774.
It inlines method GCTracer::Scope::Name so that the calculation of the
name of the trace event can be performed at compile time and optimized
away, at most call sites.

This is a reland of 370cae1d8fdddd3d7351856b861e4eef5ac15c71
which was reviewed here: https://crrev.com/c/3602511

Bug: chromium:1318062
Change-Id: I617fcad07448ebbd63790600a071e51964baf85c