Reland "[heap] Fix bogus object size computation"
This is a reland of commit 445190bf99676901431ead6451f4b30d2abaae64

The fix addresses the issue that the sweeping method didn't recompute
the object size of the space. However, since this requires the object
size, we have to do this before starting sweeing of the other spaces.
This was the original bug in first place. In order to make pointers
updating work, we have to flip the page flags for young pages.

Original change's description:
> [heap] Fix bogus object size computation
>
> The map of an object may be gone by the time we try to compute its
> size for accounting purposes.
>
> Bug: chromium:1319217
> Change-Id: I93cca766a8cedebf4ed30a3a65fd6eff5bc72bcf
> Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/3605817
> Reviewed-by: Dominik Inführ <dinfuehr@chromium.org>
> Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#80271}

Bug: chromium:1319217
Change-Id: I8d032edf96a4bf4b0faa4bbd9b0be247051c49fb