[gdb] Add a frame decorator which introspects V8 frames
Add a new `_v8_internal_Get_Frame_Name` method which returns some V8
frame information for a given frame pointer, and a corresponding gdb
FrameDecorator which calls this function, parses its output, and uses it
to decorate the backtrace.

This seemed a bit crashy on gdb < 11, so gate it behind that version.

Change-Id: I45de1f6bad75d9dcc9ba0243fa90f7107bea836a