[wasm] Notify the StreamingDecoder earlier of the NativeModule
The StreamingDecoder has a `NotifyNativeModuleCreated` function that is
used to allow the StreamingDecoder to set up the caching callback. This
function got called in `FinishCompile` so that the callback can be set
up both after the module got compiled and after the module got loaded
from the cache.

The problem with calling `NotifyNativeModuleCreated` in `FinishCompile`
is that when the module gets compiled with `--no-liftoff`, the caching
callback only gets registered after all TurboFan compilation is
finished. This was not a problem before, because without dynamic tiering
the top-tier-finished event is triggered for a callback even when the
callback gets registered after the event. However, with dynamic tiering
caching happens after the CompilationChunkFinished event, and that one
does not get triggered when a new callback gets registered.

With this CL `NotifyNativeModuleCreated` gets called right
after the NativeModule has been created.

R=clemensb@chromium.org

Change-Id: Ice3036812514b3a4c8e423aba30960157442a4b0