[cfi] Add RestartFrameTrampoline to allowed builtins list
When we want to restart an optimized frame, we have to go through
the deoptimizer. The deoptimzer has to materialize all the frames
up to, and including the frame we want to restart. Then we jump
from the deoptimizer into the RestartFrameTrampoline that
re-invokes the function.

The control flow integrity mechanism has an allow list of
builtins that the deoptimizer can jump into. RestartFrameTrampoline
is currently missing.

R=fgm@chromium.org

Bug: chromium:1303521
Change-Id: I64dc6a56e4b91642582a893ae28f3a9e199c5836