Fix rehashing of script compilation cache
The script compilation cache contains weak pointers to Script objects as
its keys. When doing a rehashing operation, any hash table needs the
ability to get the hash code for every entry in the table. However, if
the weak pointer was cleared from a key, there is no longer any way to
get the hash code for that entry.

This could be solved in a variety of ways, such as:

1. Delete all entries whose keys contain cleared weak pointers prior to
   any operation which requires rehashing. This is what I intended to
   do in https://crrev.com/c/3597106 , but the implementation had a bug.
2. Return a bogus hash code for entries whose keys contain cleared weak
   pointers. This is simple but risks poor distribution of data after
   rehashing.
3. Implement custom rehashing which can remove items as it goes, rather
   than reusing the rehashing implementation from HashTable.
4. Include the hash value in every key, so a consistent hash value is
   available even after the weak Script pointer has been cleared.

The fourth option sounds like the lowest risk to me, so this change
implements that option.

Bug: v8:12808
Change-Id: I6b19b9c8af67dcfc31b74842ba581dd141e18845