[wasm] Avoid HandleScope when triggering tier-up
Triggering tier-up can happen very often, so the runtime function should
be as slim as possible.
This CL separates the type feedback processing from triggering tier-up,
to make it easier to verify that both don't do any allocations.

It adds two DisallowGarbageCollection scopes and removes a HandleScope
which was unnecessarily created.

R=jkummerow@chromium.org

Bug: v8:12281
Change-Id: I43e7f2b449630856ac8dfb36d294fbd29191d0eb