[arm64] Fix CFI issue with short builtins
The allowlist used for `Deoptimizer::IsValidReturnAddress` depends on
fixed embedded builtin addresses. Pass a pointer to the actual builtin
code start (which may have been remapped) to this method, so that it
can calculate the offset from the start of the builtins code and
check if the return address is in the allowlist.

After this change, do not disable short builtins when CFI is enabled.

There's an important TODO for this change:
Since the builtin code pointer that's used to check whether a return address
is allowed is now writable, we will need to use pointer authentication
to sign and authenticate it.

Bug: v8:10026
Change-Id: Iafd31d3ad7e10cb17faf33e76e78d3df36edeefd