Ignore exclude_imports gn flag to avoid warning
|exclude_imports| flag is set in some of the perfetto's
proto_library targets to indicate that we don't need to generate
the proto-descriptor for the protos included in those `x.proto`
files. This flag is unused in v8 because we are using very old
protoc in v8, and it doesn't have support for `--include_imports`
command line option. In future when we upgrade protoc, we can use
this flag, similar to the CL (https://crrev.com/c/2632759) to pass
it to protoc.

Change-Id: I0689003978096798d1e966ec8485cd6af7237804