[maglev] Support property array stores
Change StoreField to StoreTaggedField, which, similar to the move of
LoadField to LoadTaggedField, operates on an offset and not a full Smi
handler. Then, add support for stores to a property array by emitting a
LoadTaggedField of the property array.

As a drive-by, fix support for const fields and HeapObject fields with
a class field type.

Bug: v8:7700
Change-Id: Iff1fec35b82d3999ff273b069e9935166f43b98f